Add saving logic to ProfileSection
A few things are proved out as a part of this commit. One example is that the TimeElapsed
profile value type now uses a DayJS duration as the deserialized value, making it
impossible to pass through deserialized values from API endpoints or gSSP functions.
However, deserializing on-the-fly has shown to be comparably ergonomic and more
consistent with other deserialization use cases in the app (as in ValueHistoryView).
.current in IProfileFieldBuilt values now refer to the ProfileField entity and not
a deserialized value.

Another thing that has been proven out is that serialized values are the ideal form
to use not only in the underlying table rows but as the value in transit through
the API as well. Serialized values prove to be quite useful for on-the-fly validation
(as in the new /api/profileFields endpoint for creating many fields).

Additionally, all inputs for Overview, Physical Wellness, and Highlights are now
using the new ProfileSection/ProfileFieldCell structure for dynamically building
out profiles based on value types.

This commit still leaves the codebase in a state that does not build. Additional
adjustments must be made to clean out hard-coded references to ProfileFieldKeys,
some of which already no longer exist.