Groundwork on building up player creation forms from existing constants
This change replaces the sub-pages in players/playerForm/* with one page that catches
any category in [profileCategory] and renders the form with the correct fields and editors
based on the information defined in interfaces/user.

More work needs to be done to re-wire this page to the little-state-machine that saves
the form's state and to the endpoint that saves the information at the end.